btc_tracker/requirements.txt
finnhub-python
ta
schedule
python-dotenv
numpy
